Room Transitions
Added room transitions.
Each room must have a room manager. The room manager has a portal point which needs to be repositioned somewhere in the level where you want the portal to spawn. The room manager has a list of enemies that are currently in the room. These enemies need to be manually added into the inspector at the moment. 
Once all enemies are dead, the portal spawns. The portal will load a random scene. This will be changed to have a preset random pattern later.